+++ .golangci.yml
@@ -1,65 +1,17 @@
 run:
-  # The default concurrency value is the number of available CPU.
-  # concurrency: 4
-  # Timeout for analysis, e.g. 30s, 5m.
-  # Default: 1m
   timeout: 5m
-  # Exit code when at least one issue was found.
-  # Default: 1
   issues-exit-code: 1
-  # Include test files or not.
-  # Default: true
   tests: true
-  # Which files to skip: they will be analyzed, but issues from them won't be reported.
-  # Default value is empty list,
-  # but there is no need to include all autogenerated files,
-  # we confidently recognize autogenerated files.
-  # If it's not please let us know.
-  # "/" will be replaced by current OS file path separator to properly work on Windows.
   skip-files:
     - ".*\\.my\\.go$"
     - lib/bad.go
-  # If set we pass it to "go list -mod={option}". From "go help modules":
-  # If invoked with -mod=readonly, the go command is disallowed from the implicit
-  # automatic updating of go.mod described above. Instead, it fails when any changes
-  # to go.mod are needed. This setting is most useful to check that go.mod does
-  # not need updates, such as in a continuous integration and testing system.
-  # If invoked with -mod=vendor, the go command assumes that the vendor
-  # directory holds the correct copies of dependencies and ignores
-  # the dependency descriptions in go.mod.
-  #
-  # Allowed values: readonly|vendor|mod
-  # By default, it isn't set.
   modules-download-mode: readonly
-  # Allow multiple parallel golangci-lint instances running.
-  # If false (default) - golangci-lint acquires file lock on start.
   allow-parallel-runners: false
-  # Define the Go version limit.
-  # Mainly related to generics support since go1.18.
-  # Default: use Go version from the go.mod file, fallback on the env var `GOVERSION`, fallback on 1.18
   go: '1.19'
 
 output:
-  # Format: colored-line-number|line-number|json|tab|checkstyle|code-climate|junit-xml|github-actions
-  #
-  # Multiple can be specified by separating them by comma, output can be provided
-  # for each of them by separating format name and path by colon symbol.
-  # Output path can be either `stdout`, `stderr` or path to the file to write to.
-  # Example: "checkstyle:report.json,colored-line-number"
-  #
-  # Default: colored-line-number
-  # format: json
-  # Print lines of code with issue.
-  # Default: true
   print-issued-lines: true
-  # Print linter name in the end of issue text.
-  # Default: true
   print-linter-name: true
-  # Make issues output unique by line.
-  # Default: true
   uniq-by-line: true
-  # Add a prefix to the output file references.
-  # Default is no prefix.
   path-prefix: ""
-  # Sort results by: filepath, line and column.
   sort-results: true
